# services/monolith/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Kvasir API
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value
...
paths:
...
  /{podId}/slices/{sliceId}/stream:
    get:
      tags:
      - Stream Api
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      - name: sliceId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: {}
  /{podId}/stream:
    get:
      tags:
      - Stream Api
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChangeRecords"
...
components:
  schemas:
...
    ChangeRecords:
      required:
      - context
      - id
      - timestamp
      type: object
      properties:
        context:
          type: object
          additionalProperties: {}
        id:
          type: string
        timestamp:
          $ref: "#/components/schemas/Instant"
        deleted: {}
        inserted: {}
...
    Instant:
      format: date-time
      type: string
      example: 2022-03-10T16:15:50Z
...